start improving printer logs
