+++ client/js/services/printer-power.service.js
