+++ server/views/partials/PrinterManager/printerLogs.modal.ejs
@@ -1,27 +1,28 @@
 <div class="col-12 col-sm-12 col-md-3 col-lg-2 col-xl-2">
     <div class="list-group" id="list-tab" role="tablist">
-        <a class="list-group-item list-group-item-action active" id="system-logs-list"
-           data-toggle="list" href="#system-logs" role="tab" aria-controls="panelview"><i
-                    class="far fa-user"></i> OctoFarm <span id="logCount">(<i
+        <a class="list-group-item list-group-item-action active notyet" id="system-logs-list"
+           data-toggle="list" href="#system-logs" role="tab" aria-controls="panelview"><i class="fa-solid fa-satellite-dish"></i> Connection Logs <span id="logCount">(<i
                         class="fas fa-spinner fa-spin"></i>)</span></a>
-        <a class="list-group-item list-group-item-action" id="system-error-list"
+        <a class="list-group-item list-group-item-action notyet" id="system-error-list"
            data-toggle="list" href="#system-error" role="tab" aria-controls="listview"><i
                     class="fas fa-exclamation-triangle"></i> Printer Errors <span
                     id="errorCount">(<i class="fas fa-spinner fa-spin"></i>)</span></a>
-        <a class="list-group-item list-group-item-action" id="system-temperature-list"
+        <a class="list-group-item list-group-item-action notyet" id="system-temperature-list"
            data-toggle="list" href="#system-temperature" role="tab" aria-controls="listview"><i
                     class="fas fa-chart-line"></i> Printer Temperature <span id="tempCount">(<i
                         class="fas fa-spinner fa-spin"></i>)</span></a>
-        <a class="list-group-item list-group-item-action" id="system-octologs-list"
-           data-toggle="list" href="#system-octologs" role="tab" aria-controls="listview"><i
-                    class="fas fa-print"></i> OctoPrints Logs <span id="octoPrintCount">(<i
-                        class="fas fa-spinner fa-spin"></i>)</span></a>
-        <a class="list-group-item list-group-item-action" id="system-octoprint-list"
+        <a class="list-group-item list-group-item-action notyet" id="system-octologs-list"
+           data-toggle="list" href="#system-octologs" role="tab" aria-controls="listview"><i class="fa-brands fa-octopus-deploy"></i> OctoPrints Logs <span id="octoPrintCount">(<i
+                        class="fas fa-spinner fa-spin"></i>)</span></adisabled>
+        <a class="list-group-item list-group-item-action notyet d-none" id="system-octoprint-list"
            data-toggle="list" href="#system-octoprint" role="tab" aria-controls="listview"><i
                     class="fas fa-plug"></i> OctoPrints Plugin Logs <span id="octoCount">(<i
                         class="fas fa-spinner fa-spin"></i>)</span></a>
-        <button class="btn btn-success" id="system-refresh-list" data-toggle="list" role="tab"
-                aria-controls="listview"><i class="fas fa-sync"></i> Refresh
+        <a class="list-group-item list-group-item-action notyet d-none" id="system-octoprint-klipper-list"
+           data-toggle="list" href="#system-octoprint" role="tab" aria-controls="listview"><i class="fa-solid fa-scissors"></i> Klipper Logs <span id="octoCount">(<i
+                        class="fas fa-spinner fa-spin"></i>)</span></a>
+        <button disabled class="btn btn-success" id="system-refresh-list" data-toggle="list" role="tab"
+                aria-controls="listview"><i class="fa-solid fa-scissors"></i> Refresh
         </button>
     </div>
 </div>
@@ -35,7 +36,6 @@
                         <thead>
                         <tr>
                             <th class=" sticky-table table-dark text-wrap">Date</th>
-                            <th scope="col" class=" sticky-table table-dark">Printer URL</th>
                             <th scope="col" class=" sticky-table table-dark">Message</th>
                         </tr>
                         </thead>
@@ -118,6 +118,27 @@
                         <!--                                                    </tr>-->
                         <!--                                                    </thead>-->
                         <tbody id="octologsLogsRows"
+                               style="height:100%; overflow-y:auto; font-family: monospace;">
+                        </tbody>
+                    </table>
+                </div>
+            </div>
+
+        </div>
+        <div class="tab-pane fade" id="system-klipperlogs" role="tabpanel"
+             aria-labelledby="server-octologs">
+            <div class="row  table-responsive-sm table-responsive-md">
+                <div class="col-12 col-sm-12 col-md-12 col-lg-12 col-xl-12">
+                    <table id="klipperLogsTable" class="table table-dark text-center">
+                        <thead>
+                        <tr>
+                            <th class=" sticky-table table-dark text-wrap">Date</th>
+                            <th scope="col" class=" sticky-table table-dark">Printer URL</th>
+                            <th scope="col" class=" sticky-table table-dark">Plugin</th>
+                            <th scope="col" class=" sticky-table table-dark">Message</th>
+                        </tr>
+                        </thead>
+                        <tbody id="klipperLogsTableRows"
                                style="height:100%; overflow-y:auto; font-family: monospace;">
                         </tbody>
                     </table>
